requirements.txt
anthropic>=0.40.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
PyGithub>=2.1.0
